Fix a few test by move files to `components` catalog
# Test Suites: 4 failed, 1 skipped, 476 passed, 480 of 481 total
# Tests:       1 failed, 24 skipped, 1 todo, 2329 passed, 2355 total
# Snapshots:   112 passed, 112 total
# Time:        303.491 s
