+++ packages/app/jest.config.renderer.js
@@ -6,8 +6,10 @@ module.exports = {
     "**/components/**/?(*.)+(test|spec).[jt]s?(x)",
     "**/actions/**/?(*.)+(test|spec).[jt]s?(x)",
     "<rootDir>/src/contact-support/**/?(*.)+(test|spec).[jt]s?(x)",
+    "<rootDir>/src/context-menu/**/?(*.)+(test|spec).[jt]s?(x)",
     "<rootDir>/src/__deprecated__/renderer/**/?(*.)+(test|spec).[jt]s?(x)",
     "<rootDir>/src/contacts/helpers/**/?(*.)+(test|spec).[jt]s?(x)",
+    "<rootDir>/src/analytic-data-tracker/helpers/**/?(*.)+(test|spec).[jt]s?(x)",
   ],
   testEnvironment: "jsdom"
 }
