+++ packages/app/src/__deprecated__/recovery-mode/components/recovery-mode.styled.tsx
