+++ packages/app/src/connecting/components/connecting.test.tsx
@@ -8,7 +8,7 @@ import { act } from "@testing-library/react"
 import { renderWithThemeAndIntl } from "App/__deprecated__/renderer/utils/render-with-theme-and-intl"
 import Connecting from "App/connecting/components/connecting.component"
 import { ErrorConnectingModalTestIds } from "App/connecting/components/error-connecting-modal-test-ids.enum"
-import { PasscodeModalTestIds } from "App/__deprecated__/passcode-modal/passcode-modal-test-ids.enum"
+import { PasscodeModalTestIds } from "App/__deprecated__/passcode-modal/components/passcode-modal-test-ids.enum"
 import { ErrorSyncModalTestIds } from "App/connecting/components/error-sync-modal/error-sync-modal-test-ids.enum"
 import { SynchronizationState } from "App/data-sync/reducers"
 import { DeviceType } from "App/device/constants"
