+++ packages/app/src/crash-dump/index.ts
@@ -4,5 +4,5 @@
  */
 
 export * from "./actions"
-export * from "./containers"
+export * from "./components"
 export * from "./reducers"
