+++ packages/app/tsconfig.json
@@ -90,7 +90,7 @@
     /* List of folders to include type definitions from. */
     // "typeRoots": [],
     /* Type declaration files to be included in compilation. */
-    "types": ["jest", "node"],
+    "types": ["jest", "node", "@testing-library/jest-dom"],
     /* Allow default imports from modules with no default export. This does not affect code emit, just typechecking. */
     // "allowSyntheticDefaultImports": true,
     /* Do not resolve the real path of symlinks. */
